requirements.txt
torch>=2.0
diffusers>=0.19.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors
Pillow
